LevelDesign
Added prefabs and cave dome/tunnel to the level design scene. Mainly to test how everything was coming together and what needs to be changed and added. Lighting still needs to be tested and worked on to an extent and cave can be made smaller. Different features need to be added and changed for the dome. Also planning on reworking the layout of the loopables and "pallets".