Completed Level Design
FULLY FINISHED DESIGN OF THE LEVEL . The only thing that needs to be added are the "Loopables" and "Pallets". Other than that everything is fully complete. I remained in the Level Design scene the entire time, so if anything conflicts it may have to do with assets because i added and deleted many things when designing.